jangoadmin/requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
boto3==1.13.15
botocore==1.16.15
certifi==2020.4.5.1
Django==3.0.6
django-environ==0.4.5
django-filter==2.4.0
django-graphiql==0.4.4
django-graphql-jwt==0.1.6
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
mysqlclient==1.4.6
pkg-resources==0.0.0
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
Rx==1.6.1
s3transfer==0.3.3
schedule==0.6.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
